feat(menu-status): Move menu icons out of svg components folder into menu widget
