+++ packages/pancake-uikit/src/components/Svg/index.tsx
@@ -22,7 +22,6 @@ export { default as ChevronUpIcon } from "./Icons/ChevronUp";
 export { default as CloseIcon } from "./Icons/Close";
 export { default as CogIcon } from "./Icons/Cog";
 export { default as CommunityIcon } from "./Icons/Community";
-export { default as Competition } from "./Icons/Competition";
 export { default as CopyIcon } from "./Icons/Copy";
 export { default as ErrorIcon } from "./Icons/Error";
 export { default as HelpIcon } from "./Icons/Help";
@@ -40,7 +39,6 @@ export { default as OpenNewIcon } from "./Icons/OpenNew";
 export { default as PancakesIcon } from "./Icons/Pancakes";
 export { default as PancakeRoundIcon } from "./Icons/PancakeRound";
 export { default as PlayCircleOutlineIcon } from "./Icons/PlayCircleOutline";
-export { default as Predictions } from "./Icons/Predictions";
 export { default as PrizeIcon } from "./Icons/Prize";
 export { default as RemoveIcon } from "./Icons/Remove";
 export { default as VerifiedIcon } from "./Icons/Verified";
