+++ packages/pancake-uikit/src/widgets/Menu/icons/TeamBattle.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Svg from "../Svg";
-import { SvgProps } from "../types";
+import Svg from "../../../components/Svg/Svg";
+import { SvgProps } from "../../../components/Svg/types";
 
 const Icon: React.FC<SvgProps> = (props) => {
   return (
