+++ packages/pancake-uikit/src/widgets/Menu/icons/index.ts
@@ -11,7 +11,9 @@ export { default as MoonIcon } from "./Moon";
 export { default as MoreIcon } from "./More";
 export { default as NftIcon } from "./Nft";
 export { default as PoolIcon } from "./Pool";
+export { default as PredictionsIcon } from "./Predictions";
 export { default as SunIcon } from "./Sun";
+export { default as TeamBattleIcon } from "./TeamBattle";
 export { default as TelegramIcon } from "./Telegram";
 export { default as TicketIcon } from "./Ticket";
 export { default as TradeIcon } from "./Trade";
